Update StdSudokuGrid class
Added in declaration of variables to be used. A 2d array for the sudoku grid. An arraylist to store the lines read from file. Int variables to track the size of arraylist and dimensions of the 2d array.

Implemented initGrid method. This method reads in a file using BufferedReader and splits the read line accordingly to grid coordinates and value which it then uses to create the sudoku grid in the 2d array.

Implemented toString method. Using two simple for loops to create a string to be printed out.