Implement KillerSudokuGrid and KillerBackTrackingSolver
KillerSudokuGrid is essentially the same as StdSudokuGrid. The only
differences are the input graphs and reading in file where
KillerSudokuGrid start with a grid of 0s and reads in instructions for
the cage restrictions, and an extra check to see if the cages are add up
to the correct amount specified. It will back track if it does not.

KillerBackTrackingSolver is the same as the standard BackTrackingSolver.